requirements.txt
# Core dependencies
torch>=2.0.0
torchaudio>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
datasets>=2.12.0

# Audio processing
faster-whisper>=0.9.0
pyannote.audio>=3.1.0
librosa>=0.10.0
soundfile>=0.12.0

# ML inference
llama-cpp-python>=0.2.0
ultralytics>=8.0.0

# API and web interface
fastapi>=0.100.0
uvicorn>=0.22.0
gradio>=3.35.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.31.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0 